Move EnumIdLookup upper level
and move it to uk.org.squeezer.framework
